Upgrade for zeitwork
This gem needs a change with the advent of zeitwork: https://github.com/fxn/zeitwerk

In particular, it should use the `config.to_prepare` method to add helpers to controllers, instead of an initializer.

See this for more information:

https://github.com/tansengming/stripe-rails/issues/168
https://github.com/tansengming/stripe-rails/pull/220/files#diff-2778b83297234370f880f1cce74b0926a940c93a09d5fe5d40c196d730be0ce9R82

Everything else in this change is adding tests and frozen_string_literal.